phone auth added to redux
